move common from proxy agent to shared
