+++ proxy_agent/src/host_clients/imds_client.rs
@@ -21,9 +21,9 @@
 //! ```
 
 use super::instance_info::InstanceInfo;
-use crate::common::{error::Error, hyper_client, logger, result::Result};
 use crate::shared_state::key_keeper_wrapper::KeyKeeperSharedState;
 use hyper::Uri;
+use proxy_agent_shared::common::{error::Error, hyper_client, logger, result::Result};
 use std::collections::HashMap;
 
 pub struct ImdsClient {
