+++ proxy_agent/src/redirector/linux.rs
@@ -2,12 +2,6 @@
 // SPDX-License-Identifier: MIT
 mod ebpf_obj;
 
-use crate::common::{
-    config, constants,
-    error::{BpfErrorType, Error},
-    logger,
-    result::Result,
-};
 use crate::redirector::{ip_to_string, AuditEntry};
 use crate::shared_state::redirector_wrapper::RedirectorSharedState;
 use aya::programs::{CgroupSockAddr, KProbe};
@@ -18,6 +12,12 @@ use aya::{
 use aya::{Btf, Ebpf, EbpfLoader};
 use ebpf_obj::{
     destination_entry, sock_addr_audit_entry, sock_addr_audit_key, sock_addr_skip_process_entry,
+};
+use proxy_agent_shared::common::{
+    config, constants,
+    error::{BpfErrorType, Error},
+    logger,
+    result::Result,
 };
 use proxy_agent_shared::telemetry::event_logger;
 use proxy_agent_shared::{logger::LoggerLevel, misc_helpers};
@@ -476,11 +476,11 @@ pub async fn update_hostga_redirect_policy(
 #[cfg(test)]
 #[cfg(feature = "test-with-root")]
 mod tests {
-    use crate::common::config;
-    use crate::common::constants;
     use crate::redirector::linux::ebpf_obj::sock_addr_audit_entry;
     use crate::redirector::linux::ebpf_obj::sock_addr_audit_key;
     use aya::maps::HashMap;
+    use proxy_agent_shared::common::config;
+    use proxy_agent_shared::common::constants;
     use proxy_agent_shared::misc_helpers;
     use std::env;
