+++ proxy_agent_shared/src/common.rs
