+++ proxy_agent_shared/src/common/error.rs
