+++ proxy_agent_shared/src/common/result.rs
